Refactor base_node into modular package with deprecation support
